ml_server_app/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
Pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
pdf2image>=1.16.3
spacy>=3.7.0
scikit-learn>=1.3.0
tensorflow>=2.15.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0
python-dateutil>=2.8.2
tqdm>=4.66.0
requests>=2.31.0